feat(skill): added skill proficiency and upgrade system for levels above 7
- Extended skill level cap to 20
- Proficiency increases based on valid usage count
- Automatic level-up upon filling the progress bar
- Damage and stat bonuses applied based on per-level configuration
- Retained visual effects from level 7 skills
- Added skill proficiency management page to the Admin Menu
- Supports global, per-skill, and per-level configurations
- Client patch to display proficiency bars for skills level 7 and above
- Added tools to check, apply, and revert the client patch

+++ src/nro/models/server/Manager.java
@@ -460,7 +460,7 @@ public final class Manager {
                     skill.dx = Integer.parseInt(String.valueOf(dts.get("dx")));
                     skill.dy = Integer.parseInt(String.valueOf(dts.get("dy")));
                     skill.maxFight = Integer.parseInt(String.valueOf(dts.get("max_fight")));
-                    skill.damage = Short.parseShort(String.valueOf(dts.get("damage")));
+                    skill.damage = Integer.parseInt(String.valueOf(dts.get("damage")));
                     skill.price = Short.parseShort(String.valueOf(dts.get("price")));
                     skill.moreInfo = String.valueOf(dts.get("info"));
                     skillTemplate.skillss.add(skill);
